Persistence API - Allow replaying of events within a range
There are use cases (mainly debugging / testing) where it would be useful to be able to replay up to a certain point in time or a certain index. For example, if we want to replay until just before something happened (i.e. unexpected behavior of the system, bug, crash, "the system went down at 12:00" etc..) then apply some messages and observe what happens.

closes #188