Use an optimized version Rendezvous hashing for PartitionActivator.
Before this, balance of Actors between nodes were not uniform (up to 40% distribution difference in a usual deployment, using Guids as identity), this makes it very uniform in comparison (down to less than 1%).
It performs as good as or better than MemberHashRing (in use by PartitionIdentity) up to 30 nodes, and comparable up to 100 nodes.